refactor(tools): split the single-file crate by concern
leviath-tools was one 2,601-line lib.rs with a 940-line impl block. The
concerns now live in their own modules - platform capabilities, the
sandbox context and alias table, tool definitions, and tool execution -
with the definition and execution halves as separate impl blocks on the
same BuiltinTools, so every call site stays unchanged. lib.rs keeps the
struct, its constructors, and the crate-facing re-exports.

Pure moves; the suite and the coverage gate are green on both sides.
