+++ crates/leviath-tools/src/lib.rs
@@ -11,216 +11,14 @@ use std::sync::{Arc, Mutex, PoisonError};
 use tokio::process::Command;
 use tokio::time::{Duration, timeout};
 
-/// Put `cmd` in its own process group, so the whole command tree can be
-/// signalled as one. `leviath_sys::configure_detached` does this for a
-/// `std::process::Command`; the tool lane uses tokio's, which has its own
-/// (Unix-only) setter. A no-op where the platform has no process groups —
-/// there, killing the direct child is all that exists.
-pub fn own_process_group(cmd: &mut Command) {
-    #[cfg(unix)]
-    cmd.process_group(0);
-    #[cfg(not(unix))]
-    let _ = cmd;
-}
-
-/// SIGKILLs a shell's whole process group when dropped.
-///
-/// `kill_on_drop` reaps the shell itself; this reaps what the shell started.
-/// Held for the duration of one shell tool call, so it fires on every exit path
-/// — normal completion (where the group is already gone and the signal is a
-/// harmless no-op), timeout, and the future being dropped because the agent was
-/// cancelled.
-pub struct ProcessGroupReaper(pub u32);
-
-impl Drop for ProcessGroupReaper {
-    fn drop(&mut self) {
-        // The group is usually already gone; failing to signal it is expected.
-        let _ = leviath_sys::kill_process_group(self.0);
-    }
-}
-
-/// A platform feature a built-in tool depends on.
-///
-/// Each built-in declares the capabilities it requires (see
-/// [`tool_required_capabilities`]); the current platform declares what it
-/// provides (see [`PlatformCapabilities`]). A tool whose requirements aren't
-/// met by the platform simply doesn't register — it's dropped from
-/// advertisement, name-recognition, and dispatch. This lets platform-specific
-/// tools (like `shell`, which spawns OS processes) coexist without per-tool
-/// `#[cfg]` gates or one-off boolean flags.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub enum ToolCapability {
-    /// Can launch child processes (e.g. the `shell` tool, stdio MCP servers).
-    ProcessSpawn,
-    /// Can read and write files.
-    FileSystem,
-    /// Can make outbound network/HTTP requests.
-    Network,
-}
-
-/// The set of [`ToolCapability`]s the current platform provides.
-///
-/// Detection is compile-time: [`current`](Self::current) resolves to the
-/// capability set for the build target. Desktop targets provide everything; a
-/// future mobile or wasm host would provide a reduced set (no
-/// [`ProcessSpawn`](ToolCapability::ProcessSpawn)). This mirrors the
-/// `leviath-sys` crate's approach of centralizing platform `#[cfg]` branching
-/// in one place rather than scattering it across call sites.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PlatformCapabilities {
-    capabilities: HashSet<ToolCapability>,
-}
-
-impl PlatformCapabilities {
-    /// Capabilities for a desktop host: process spawning, filesystem, network.
-    pub fn desktop() -> Self {
-        Self {
-            capabilities: HashSet::from([
-                ToolCapability::ProcessSpawn,
-                ToolCapability::FileSystem,
-                ToolCapability::Network,
-            ]),
-        }
-    }
-
-    /// Capabilities for a mobile host: filesystem and network, but no process
-    /// spawning (so the `shell` tool doesn't register).
-    pub fn mobile() -> Self {
-        Self {
-            capabilities: HashSet::from([ToolCapability::FileSystem, ToolCapability::Network]),
-        }
-    }
-
-    /// The capability set for the platform this binary was built for.
-    ///
-    /// Only desktop targets are built today, so this resolves to
-    /// [`desktop`](Self::desktop). A future mobile/wasm target adds a `cfg` arm
-    /// here returning [`mobile`](Self::mobile) — every built-in then auto-filters
-    /// against it with no other change.
-    pub fn current() -> Self {
-        Self::desktop()
-    }
-
-    /// Build a capability set from an explicit collection (tests, future hosts).
-    pub fn from_capabilities(caps: impl IntoIterator<Item = ToolCapability>) -> Self {
-        Self {
-            capabilities: caps.into_iter().collect(),
-        }
-    }
-
-    /// Whether the platform provides `cap`.
-    pub fn supports(&self, cap: ToolCapability) -> bool {
-        self.capabilities.contains(&cap)
-    }
-
-    /// Whether the platform provides *every* capability in `required`. An empty
-    /// requirement slice is always satisfied.
-    pub fn satisfies(&self, required: &[ToolCapability]) -> bool {
-        required.iter().all(|c| self.supports(*c))
-    }
-}
-
-impl Default for PlatformCapabilities {
-    fn default() -> Self {
-        Self::current()
-    }
-}
-
-/// The capabilities a built-in tool requires to function.
-///
-/// Keyed by *canonical* tool name (resolve aliases via [`canonical_tool_name`]
-/// first). An empty slice means the tool is platform-agnostic and always
-/// available — this covers the runtime-handled tools (`present_for_review`,
-/// `ask_user_*`, `edit_document`, `context_*`) which touch neither the OS
-/// process table nor the filesystem directly.
-pub fn tool_required_capabilities(canonical_name: &str) -> &'static [ToolCapability] {
-    match canonical_name {
-        "shell" => &[ToolCapability::ProcessSpawn],
-        "read_file" | "read_files" | "write_file" | "edit_file" | "list_dir" => {
-            &[ToolCapability::FileSystem]
-        }
-        _ => &[],
-    }
-}
-
-/// Context for tool execution — defines the sandbox root.
-pub struct ToolContext {
-    /// Absolute working directory. All file operations are confined here.
-    pub workdir: PathBuf,
-    /// Per-path advisory locks serializing concurrent mutating file operations
-    /// (`write_file`/`edit_file`) on the *same* file. Fan-out sub-agent workers
-    /// share one process and one workdir, so an in-process lock map keyed by
-    /// canonical path is sufficient (no OS `flock` needed) to prevent lost
-    /// updates when two workers touch the same file. Different files never
-    /// contend.
-    file_locks: Arc<Mutex<HashMap<PathBuf, Arc<tokio::sync::Mutex<()>>>>>,
-}
-
-impl ToolContext {
-    /// Create a new context. Attempts to canonicalize the working directory.
-    pub fn new(workdir: PathBuf) -> Self {
-        let workdir = std::fs::canonicalize(&workdir).unwrap_or(workdir);
-        Self {
-            workdir,
-            file_locks: Arc::new(Mutex::new(HashMap::new())),
-        }
-    }
-
-    /// Get (or create) the advisory lock for `path`. The map mutex is held only
-    /// briefly to look up / insert; the returned per-file lock is what callers
-    /// `.await` on across their read-modify-write.
-    fn lock_for(&self, path: &Path) -> Arc<tokio::sync::Mutex<()>> {
-        let mut map = self
-            .file_locks
-            .lock()
-            .unwrap_or_else(PoisonError::into_inner);
-        map.entry(path.to_path_buf())
-            .or_insert_with(|| Arc::new(tokio::sync::Mutex::new(())))
-            .clone()
-    }
-}
-
-/// Alias → canonical built-in tool name.
-///
-/// A blueprint's `available_tools` may name a built-in by any alias listed here;
-/// it resolves to the canonical tool that is advertised to the model and
-/// executed. This is the single source of truth for aliases — [`names`],
-/// [`BuiltinTools::execute`], and the daemon's `available_tools` filtering all go
-/// through [`canonical_tool_name`], so adding a row here is all it takes to add
-/// an alias everywhere. Add rows only for genuine synonyms of an existing tool.
-///
-/// [`names`]: BuiltinTools::names
-pub const TOOL_ALIASES: &[(&str, &str)] = &[
-    // `bash` is the familiar name for the general shell tool.
-    ("bash", "shell"),
-];
-
-/// Resolve `name` through [`TOOL_ALIASES`] to its canonical built-in name.
-///
-/// Returns the input unchanged when it is not an alias — which includes every
-/// canonical built-in and every MCP tool name, so this is safe to apply to any
-/// tool name before matching it against a definition.
-pub fn canonical_tool_name(name: &str) -> &str {
-    for (alias, canonical) in TOOL_ALIASES {
-        if *alias == name {
-            return canonical;
-        }
-    }
-    name
-}
-
-/// Redirects shell command execution off the host into a sandbox.
-///
-/// The default (no executor) runs the command directly on the host — the exact
-/// prior behavior. An implementor (the daemon's `SandboxManager`) returns a
-/// [`tokio::process::Command`] that runs `command` inside a container or Linux
-/// namespace instead. The implementor owns any per-stage sandbox state, so the
-/// same handle is used for the agent's whole life; only shell execution is
-/// affected (file tools stay on the host, over the bind-mounted workdir).
-pub trait ShellExecutor: Send + Sync {
-    /// Build the process that runs `command` via `shell flag` for `workdir`.
-    fn build_command(&self, shell: &str, flag: &str, command: &str, workdir: &Path) -> Command;
-}
+// The tool families, one module per concern; lib.rs keeps the struct and
+// its constructors.
+mod context;
+mod defs;
+mod exec;
+mod platform;
+pub use context::*;
+pub use platform::*;
 
 /// Built-in tools: read_file, write_file, edit_file, list_dir, shell.
 ///
@@ -267,911 +65,6 @@ impl BuiltinTools {
     fn available(&self, canonical_name: &str) -> bool {
         self.platform
             .satisfies(tool_required_capabilities(canonical_name))
-    }
-
-    /// All tool definitions to advertise to the LLM, minus any whose required
-    /// platform capabilities aren't provided by the current platform.
-    pub fn tool_defs(&self) -> Vec<Tool> {
-        let mut defs = vec![
-            Tool {
-                name: "read_file".to_string(),
-                description: "Read the complete contents of a file. Use this to examine existing code, configurations, or data files before making changes.".to_string(),
-                parameters: json!({
-                    "type": "object",
-                    "properties": {
-                        "path": {
-                            "type": "string",
-                            "description": "Path to the file, relative to the working directory"
-                        }
-                    },
-                    "required": ["path"]
-                }),
-            },
-            Tool {
-                name: "write_file".to_string(),
-                description: "Write content to a file, creating it (and any parent directories) if necessary. Use this to create new files or completely replace existing file content.".to_string(),
-                parameters: json!({
-                    "type": "object",
-                    "properties": {
-                        "path": {
-                            "type": "string",
-                            "description": "Path to the file, relative to the working directory"
-                        },
-                        "content": {
-                            "type": "string",
-                            "description": "The full content to write to the file"
-                        }
-                    },
-                    "required": ["path", "content"]
-                }),
-            },
-            Tool {
-                name: "edit_file".to_string(),
-                description: "Replace an exact string in an existing file. The old_str must appear exactly once in the file. Use this for targeted edits rather than rewriting entire files.".to_string(),
-                parameters: json!({
-                    "type": "object",
-                    "properties": {
-                        "path": {
-                            "type": "string",
-                            "description": "Path to the file, relative to the working directory"
-                        },
-                        "old_str": {
-                            "type": "string",
-                            "description": "The exact string to replace. Must appear exactly once in the file."
-                        },
-                        "new_str": {
-                            "type": "string",
-                            "description": "The string to replace old_str with"
-                        }
-                    },
-                    "required": ["path", "old_str", "new_str"]
-                }),
-            },
-            Tool {
-                name: "list_dir".to_string(),
-                description: "List the contents of a directory. Use this to explore the file structure before reading or writing files.".to_string(),
-                parameters: json!({
-                    "type": "object",
-                    "properties": {
-                        "path": {
-                            "type": "string",
-                            "description": "Path to the directory, relative to the working directory. Defaults to the working directory root if omitted."
-                        }
-                    },
-                    "required": []
-                }),
-            },
-            Tool {
-                name: "read_files".to_string(),
-                description: "Read multiple files at once. Returns the contents of all requested files in a single response, separated by file path headers. More efficient than calling read_file repeatedly. Use this when you need to read several files (e.g. after list_dir).".to_string(),
-                parameters: json!({
-                    "type": "object",
-                    "properties": {
-                        "paths": {
-                            "type": "array",
-                            "items": { "type": "string" },
-                            "description": "Array of file paths relative to the working directory"
-                        }
-                    },
-                    "required": ["paths"]
-                }),
-            },
-            Tool {
-                name: "shell".to_string(),
-                description: "Execute a shell command in the working directory. Uses the system shell (bash/zsh on Unix, cmd on Windows). Use this for build commands, running tests, installing dependencies, or other shell operations. Has a 60-second timeout.".to_string(),
-                parameters: json!({
-                    "type": "object",
-                    "properties": {
-                        "command": {
-                            "type": "string",
-                            "description": "The shell command to execute"
-                        }
-                    },
-                    "required": ["command"]
-                }),
-            },
-            Tool {
-                name: "present_for_review".to_string(),
-                description: "Present a document, plan, or report to the user for review. The agent run will pause and the dashboard will display the document prominently. Use this when you want the user to read and approve something before you continue — for example, a technical design, an implementation plan, or a summary report. The user can provide feedback or simply acknowledge to continue.".to_string(),
-                parameters: json!({
-                    "type": "object",
-                    "properties": {
-                        "title": {
-                            "type": "string",
-                            "description": "Short title for the review prompt shown to the user (e.g. 'Implementation Plan Ready for Review')"
-                        },
-                        "markdown": {
-                            "type": "string",
-                            "description": "The markdown document to present to the user. Supports headings, lists, code blocks, and mermaid diagrams."
-                        }
-                    },
-                    "required": ["title", "markdown"]
-                }),
-            },
-            Tool {
-                name: "ask_user_text".to_string(),
-                description: "Ask the user a free-form question and wait for their written answer. The run pauses until they respond. Use this when you need clarification, missing information, or a specific detail only the user knows — decide for yourself when this is necessary; don't ask about things you can figure out on your own.".to_string(),
-                parameters: json!({
-                    "type": "object",
-                    "properties": {
-                        "prompt": {
-                            "type": "string",
-                            "description": "The question to ask the user"
-                        }
-                    },
-                    "required": ["prompt"]
-                }),
-            },
-            Tool {
-                name: "ask_user_choice".to_string(),
-                description: "Ask the user to pick one option from a list and wait for their answer. The run pauses until they respond. Use this when you have a small number of distinct paths forward and want the user to decide which one, rather than guessing yourself.".to_string(),
-                parameters: json!({
-                    "type": "object",
-                    "properties": {
-                        "prompt": {
-                            "type": "string",
-                            "description": "The question to ask the user"
-                        },
-                        "options": {
-                            "type": "array",
-                            "items": { "type": "string" },
-                            "description": "At least two options for the user to choose from"
-                        }
-                    },
-                    "required": ["prompt", "options"]
-                }),
-            },
-            Tool {
-                name: "ask_user_confirm".to_string(),
-                description: "Ask the user a yes/no question and wait for their answer. The run pauses until they respond. Use this for a quick go/no-go decision before doing something significant or hard to undo.".to_string(),
-                parameters: json!({
-                    "type": "object",
-                    "properties": {
-                        "prompt": {
-                            "type": "string",
-                            "description": "The yes/no question to ask the user"
-                        }
-                    },
-                    "required": ["prompt"]
-                }),
-            },
-            Tool {
-                name: "edit_document".to_string(),
-                description: "Present a document to the user in an editable field pre-filled with its current text, and wait for them to edit it directly. The run pauses until they submit. Use this when the user wants to modify content themselves (e.g. tweak a plan or draft) rather than describe changes for you to make. Pass the current full text as `content`; the returned text is the user's edited version, which you should adopt as authoritative.".to_string(),
-                parameters: json!({
-                    "type": "object",
-                    "properties": {
-                        "content": {
-                            "type": "string",
-                            "description": "The current full document text to present for editing"
-                        },
-                        "prompt": {
-                            "type": "string",
-                            "description": "Optional instruction shown above the editable field"
-                        }
-                    },
-                    "required": ["content"]
-                }),
-            },
-            Tool {
-                name: "context_write".to_string(),
-                description: "Store or update content in a named section of your context window. This content will be included in your system prompt on subsequent turns, making it available for reference. Use this to save analysis, plans, notes, or structured information. If a key is provided and an entry with that key already exists, it will be replaced with the new content.".to_string(),
-                parameters: json!({
-                    "type": "object",
-                    "properties": {
-                        "region": {
-                            "type": "string",
-                            "description": "Name of the context window section (e.g. 'architecture', 'plan')"
-                        },
-                        "key": {
-                            "type": "string",
-                            "description": "Key for the entry. Replaces existing entry with the same key."
-                        },
-                        "content": {
-                            "type": "string",
-                            "description": "Content to store"
-                        }
-                    },
-                    "required": ["region", "content"]
-                }),
-            },
-            Tool {
-                name: "context_append".to_string(),
-                description: "Add content to an existing section of your context window without replacing what's already there.".to_string(),
-                parameters: json!({
-                    "type": "object",
-                    "properties": {
-                        "region": {
-                            "type": "string",
-                            "description": "Name of the context window section"
-                        },
-                        "key": {
-                            "type": "string",
-                            "description": "Key for the entry"
-                        },
-                        "content": {
-                            "type": "string",
-                            "description": "Content to append"
-                        }
-                    },
-                    "required": ["region", "content"]
-                }),
-            },
-            Tool {
-                name: "context_read".to_string(),
-                description: "Read what's currently stored in a section of your context window. If no key is specified and the section contains keyed entries, returns a summary of all keys and their sizes.".to_string(),
-                parameters: json!({
-                    "type": "object",
-                    "properties": {
-                        "region": {
-                            "type": "string",
-                            "description": "Name of the context window section to read"
-                        },
-                        "key": {
-                            "type": "string",
-                            "description": "Key of a specific entry to read"
-                        }
-                    },
-                    "required": ["region"]
-                }),
-            },
-            Tool {
-                name: "context_delete".to_string(),
-                description: "Remove a specific keyed entry from a section of your context window.".to_string(),
-                parameters: json!({
-                    "type": "object",
-                    "properties": {
-                        "region": {
-                            "type": "string",
-                            "description": "Name of the context window section"
-                        },
-                        "key": {
-                            "type": "string",
-                            "description": "Key of the entry to remove"
-                        }
-                    },
-                    "required": ["region", "key"]
-                }),
-            },
-            Tool {
-                name: "context_list".to_string(),
-                description: "List available sections of your context window with their current usage — section names, token counts, and number of entries. Use this to see what's available and what you've already stored.".to_string(),
-                parameters: json!({
-                    "type": "object",
-                    "properties": {
-                        "region": {
-                            "type": "string",
-                            "description": "Optional region name to list keys for"
-                        }
-                    },
-                    "required": []
-                }),
-            },
-        ];
-        defs.retain(|t| self.available(&t.name));
-        defs
-    }
-
-    /// Tool definitions for sub-agent management tools.
-    ///
-    /// These are advertised to the LLM but executed externally (by the CLI's
-    /// tool registry) since they require access to the AgentEngine.
-    pub fn subagent_tool_defs() -> Vec<Tool> {
-        vec![
-            Tool {
-                name: "spawn_agent".to_string(),
-                description: "Spawn a sub-agent from a blueprint to work on a task. Returns the new agent's ID. If wait=true, blocks until the sub-agent completes and returns its result.".to_string(),
-                parameters: json!({
-                    "type": "object",
-                    "properties": {
-                        "blueprint": {
-                            "type": "string",
-                            "description": "Name of the agent blueprint to spawn"
-                        },
-                        "task": {
-                            "type": "string",
-                            "description": "Task prompt for the sub-agent"
-                        },
-                        "wait": {
-                            "type": "boolean",
-                            "description": "If true, block until the sub-agent completes and return its result. Default: false",
-                            "default": false
-                        },
-                        "seed_context": {
-                            "type": "string",
-                            "description": "Optional initial context to inject into the sub-agent's first Pinned region"
-                        },
-                        "max_child_depth": {
-                            "type": "integer",
-                            "description": "Optional max depth for the sub-agent's own children"
-                        }
-                    },
-                    "required": ["blueprint", "task"]
-                }),
-            },
-            Tool {
-                name: "check_agent".to_string(),
-                description: "Check the status of a sub-agent. Returns its current status and result if complete. Non-blocking.".to_string(),
-                parameters: json!({
-                    "type": "object",
-                    "properties": {
-                        "agent_id": {
-                            "type": "string",
-                            "description": "ID of the agent to check"
-                        }
-                    },
-                    "required": ["agent_id"]
-                }),
-            },
-            Tool {
-                name: "wait_for_agent".to_string(),
-                description: "Block until a sub-agent completes, then return its final result.".to_string(),
-                parameters: json!({
-                    "type": "object",
-                    "properties": {
-                        "agent_id": {
-                            "type": "string",
-                            "description": "ID of the agent to wait for"
-                        }
-                    },
-                    "required": ["agent_id"]
-                }),
-            },
-            Tool {
-                name: "send_to_agent".to_string(),
-                description: "Send a message to a running sub-agent's context window.".to_string(),
-                parameters: json!({
-                    "type": "object",
-                    "properties": {
-                        "agent_id": {
-                            "type": "string",
-                            "description": "ID of the target agent"
-                        },
-                        "message": {
-                            "type": "string",
-                            "description": "Message content to send"
-                        },
-                        "target_region": {
-                            "type": "string",
-                            "description": "Context region to deliver to (default: conversation)"
-                        }
-                    },
-                    "required": ["agent_id", "message"]
-                }),
-            },
-            Tool {
-                name: "kill_agent".to_string(),
-                description: "Kill a sub-agent and all its descendants. Sets their cancellation tokens and marks them as cancelled.".to_string(),
-                parameters: json!({
-                    "type": "object",
-                    "properties": {
-                        "agent_id": {
-                            "type": "string",
-                            "description": "ID of the agent to kill"
-                        }
-                    },
-                    "required": ["agent_id"]
-                }),
-            },
-        ]
-    }
-
-    /// Names of sub-agent tools.
-    pub fn subagent_tool_names() -> Vec<String> {
-        vec![
-            "spawn_agent".to_string(),
-            "check_agent".to_string(),
-            "wait_for_agent".to_string(),
-            "send_to_agent".to_string(),
-            "kill_agent".to_string(),
-        ]
-    }
-
-    /// Names of all built-in tools, including every alias in [`TOOL_ALIASES`].
-    ///
-    /// Aliases are included so tool-call dispatch recognizes a call arriving
-    /// under an alias name as a built-in; the canonical names are what get
-    /// advertised to the model.
-    pub fn names(&self) -> Vec<String> {
-        let mut names: Vec<String> = [
-            "read_file",
-            "read_files",
-            "write_file",
-            "edit_file",
-            "list_dir",
-            "shell",
-            "present_for_review",
-            "ask_user_text",
-            "ask_user_choice",
-            "ask_user_confirm",
-            "edit_document",
-            "context_write",
-            "context_append",
-            "context_read",
-            "context_delete",
-            "context_list",
-        ]
-        .iter()
-        // Drop any canonical built-in the current platform can't provide, so a
-        // filtered-out tool (e.g. `shell` without `ProcessSpawn`) isn't even
-        // recognized as a built-in on dispatch.
-        .filter(|n| self.available(n))
-        .map(|s| s.to_string())
-        .collect();
-        // Include an alias only when its canonical target survived filtering
-        // (so `bash` disappears together with `shell`).
-        names.extend(
-            TOOL_ALIASES
-                .iter()
-                .filter(|(_, canonical)| self.available(canonical))
-                .map(|(alias, _)| alias.to_string()),
-        );
-        names
-    }
-
-    /// Execute a built-in tool by name (resolving aliases), returning the result
-    /// as a string.
-    pub async fn execute(&self, name: &str, args: Value) -> String {
-        let canonical = canonical_tool_name(name);
-        // A tool whose platform capabilities aren't met never advertises, but a
-        // caller could still dispatch to it directly — reject it here too.
-        if !self.available(canonical) {
-            return format!("[error] tool '{}' is not available on this platform", name);
-        }
-        match canonical {
-            "read_file" => self.read_file(&args).await,
-            "read_files" => self.read_files(&args).await,
-            "write_file" => self.write_file(&args).await,
-            "edit_file" => self.edit_file(&args).await,
-            "list_dir" => self.list_dir(&args).await,
-            "shell" => self.shell(&args).await,
-            n if n.starts_with("context_") => {
-                "[error] context tools must be handled by the runtime".to_string()
-            }
-            _ => format!("[error] Unknown built-in tool: {}", name),
-        }
-    }
-
-    /// Refuse to create anything when the working directory itself is gone.
-    ///
-    /// `write_file` calls `create_dir_all`, which would otherwise silently
-    /// resurrect a workspace an external harness deleted mid-run — leaving the
-    /// agent writing into an empty tree that no longer resembles the checkout it
-    /// reasoned about, and masking the loss from the runtime's health check
-    /// (issue #107). Creating *sub*directories inside a live workspace is
-    /// untouched; only a missing workspace root is refused.
-    fn ensure_workspace(&self) -> Result<(), String> {
-        if std::fs::metadata(&self.ctx.workdir).is_ok_and(|m| m.is_dir()) {
-            return Ok(());
-        }
-        Err(format!(
-            "[error] workspace '{}' is no longer accessible",
-            self.ctx.workdir.display()
-        ))
-    }
-
-    /// Resolve a requested path to an absolute path inside the workdir.
-    ///
-    /// Two checks, because either alone is insufficient:
-    ///
-    /// 1. **Lexical.** `..` and `.` are folded out and the result must sit under
-    ///    the workdir. Cheap, and it catches the obvious `../../etc/passwd`.
-    /// 2. **Symbolic.** The deepest *existing* ancestor is canonicalized and the
-    ///    result re-checked. Without this the containment was purely textual: a
-    ///    symlink at `<workdir>/link` pointing at `/` made
-    ///    `read_file("link/etc/passwd")` normalize to a path that starts with the
-    ///    workdir, pass, and then be followed by `fs::read_to_string`. The same
-    ///    hole let `write_file` overwrite `~/.ssh/authorized_keys`.
-    ///
-    /// That mattered most where the containment is load-bearing. Leviath's file
-    /// tools run **on the host over the bind-mounted workdir** even when the
-    /// stage's `shell` is confined to a container — so a symlink the agent
-    /// created inside the container escaped the container through the file
-    /// tools. It also matters for a freshly cloned repository, which is exactly
-    /// what a coding agent operates on and which can carry a checked-in symlink
-    /// pointing anywhere.
-    ///
-    /// The check is not TOCTOU-proof: a symlink planted between this call and the
-    /// subsequent `open` still wins. Closing that needs `openat`/`O_NOFOLLOW`
-    /// throughout, which is a larger change; this stops the planted-symlink case,
-    /// which is the one an agent can actually arrange.
-    fn resolve(&self, requested: &str) -> anyhow::Result<PathBuf> {
-        Self::resolve_within(requested, &self.ctx.workdir, resolves_within)
-    }
-
-    /// Core of [`resolve`](Self::resolve) with the containment check injected.
-    ///
-    /// A `fn` pointer (not `impl Fn`) so there is one monomorphization, matching
-    /// the seam idiom used elsewhere in the workspace. The seam exists because
-    /// the refusal cannot be reached otherwise on every platform: producing the
-    /// escape needs a real symlink, and creating one on Windows requires a
-    /// privilege CI runners do not have. Injecting the predicate lets the
-    /// refusal itself be tested everywhere, while the `#[cfg(unix)]` tests below
-    /// still prove the real filesystem behaviour end to end.
-    fn resolve_within(
-        requested: &str,
-        workdir: &Path,
-        within: fn(&Path, &Path) -> bool,
-    ) -> anyhow::Result<PathBuf> {
-        let raw = if Path::new(requested).is_absolute() {
-            PathBuf::from(requested)
-        } else {
-            workdir.join(requested)
-        };
-
-        // Normalize by resolving .. and . without requiring the path to exist.
-        let mut normalized = PathBuf::new();
-        for component in raw.components() {
-            match component {
-                Component::ParentDir => {
-                    if !normalized.pop() {
-                        anyhow::bail!("path '{}' escapes the working directory", requested);
-                    }
-                }
-                c => normalized.push(c),
-            }
-        }
-
-        if !normalized.starts_with(workdir) {
-            anyhow::bail!("path '{}' would escape the working directory", requested);
-        }
-
-        if !within(&normalized, workdir) {
-            anyhow::bail!(
-                "path '{requested}' resolves outside the working directory through a symlink"
-            );
-        }
-
-        Ok(normalized)
-    }
-
-    async fn read_file(&self, args: &Value) -> String {
-        let path_str = match args.get("path").and_then(|v| v.as_str()) {
-            Some(p) => p,
-            None => return "[error] missing 'path' argument".to_string(),
-        };
-
-        let path = match self.resolve(path_str) {
-            Ok(p) => p,
-            Err(e) => return format!("[error] {}", e),
-        };
-
-        match std::fs::read_to_string(&path) {
-            Ok(content) => content,
-            Err(e) => format!("[error] Failed to read '{}': {}", path_str, e),
-        }
-    }
-
-    async fn read_files(&self, args: &Value) -> String {
-        let paths = match args.get("paths").and_then(|v| v.as_array()) {
-            Some(arr) => arr,
-            None => return "[error] missing 'paths' argument (expected array)".to_string(),
-        };
-
-        if paths.is_empty() {
-            return "[error] 'paths' array is empty".to_string();
-        }
-
-        let mut results = Vec::with_capacity(paths.len());
-        for path_val in paths {
-            let path_str = match path_val.as_str() {
-                Some(p) => p,
-                None => {
-                    results.push("[error] non-string path in array".to_string());
-                    continue;
-                }
-            };
-
-            let path = match self.resolve(path_str) {
-                Ok(p) => p,
-                Err(e) => {
-                    results.push(format!("### [{}]\n[error] {}", path_str, e));
-                    continue;
-                }
-            };
-
-            match std::fs::read_to_string(&path) {
-                Ok(content) => {
-                    results.push(format!("### [{}]\n{}", path_str, content));
-                }
-                Err(e) => {
-                    results.push(format!("### [{}]\n[error] Failed to read: {}", path_str, e));
-                }
-            }
-        }
-
-        results.join("\n\n")
-    }
-
-    async fn write_file(&self, args: &Value) -> String {
-        let path_str = match args.get("path").and_then(|v| v.as_str()) {
-            Some(p) => p,
-            None => return "[error] missing 'path' argument".to_string(),
-        };
-        let content = match args.get("content").and_then(|v| v.as_str()) {
-            Some(c) => c,
-            None => return "[error] missing 'content' argument".to_string(),
-        };
-        if let Err(e) = self.ensure_workspace() {
-            return e;
-        }
-
-        let path = match self.resolve(path_str) {
-            Ok(p) => p,
-            Err(e) => return format!("[error] {}", e),
-        };
-
-        // Serialize concurrent writes to the same file (fan-out workers).
-        let lock = self.ctx.lock_for(&path);
-        let _guard = lock.lock().await;
-
-        let parent = {
-            let mut p = path.clone();
-            p.pop();
-            p
-        };
-        if let Err(e) = std::fs::create_dir_all(&parent) {
-            return format!(
-                "[error] Failed to create directories for '{}': {}",
-                path_str, e
-            );
-        }
-
-        match std::fs::write(&path, content) {
-            Ok(()) => format!(
-                "Successfully wrote {} bytes to '{}'",
-                content.len(),
-                path_str
-            ),
-            Err(e) => format!("[error] Failed to write '{}': {}", path_str, e),
-        }
-    }
-
-    async fn edit_file(&self, args: &Value) -> String {
-        let path_str = match args.get("path").and_then(|v| v.as_str()) {
-            Some(p) => p,
-            None => return "[error] missing 'path' argument".to_string(),
-        };
-        let old_str = match args.get("old_str").and_then(|v| v.as_str()) {
-            Some(s) => s,
-            None => return "[error] missing 'old_str' argument".to_string(),
-        };
-        let new_str = match args.get("new_str").and_then(|v| v.as_str()) {
-            Some(s) => s,
-            None => return "[error] missing 'new_str' argument".to_string(),
-        };
-        if let Err(e) = self.ensure_workspace() {
-            return e;
-        }
-
-        let path = match self.resolve(path_str) {
-            Ok(p) => p,
-            Err(e) => return format!("[error] {}", e),
-        };
-
-        // Serialize the read-modify-write against concurrent edits/writes to the
-        // same file (fan-out workers), preventing lost updates.
-        let lock = self.ctx.lock_for(&path);
-        let _guard = lock.lock().await;
-
-        let content = match std::fs::read_to_string(&path) {
-            Ok(c) => c,
-            Err(e) => return format!("[error] Failed to read '{}': {}", path_str, e),
-        };
-
-        let count = content.matches(old_str).count();
-        match count {
-            0 => format!(
-                "[error] String not found in '{}'. Ensure old_str matches the file exactly.",
-                path_str
-            ),
-            1 => {
-                let new_content = content.replacen(old_str, new_str, 1);
-                match std::fs::write(&path, &new_content) {
-                    Ok(()) => format!("Successfully edited '{}'", path_str),
-                    Err(e) => format!("[error] Failed to write '{}': {}", path_str, e),
-                }
-            }
-            n => format!(
-                "[error] Found {} occurrences of the string in '{}'. old_str must be unique.",
-                n, path_str
-            ),
-        }
-    }
-
-    async fn list_dir(&self, args: &Value) -> String {
-        let path_str = args.get("path").and_then(|v| v.as_str()).unwrap_or(".");
-
-        let path = match self.resolve(path_str) {
-            Ok(p) => p,
-            Err(e) => return format!("[error] {}", e),
-        };
-
-        let entries = match std::fs::read_dir(&path) {
-            Ok(e) => e,
-            Err(e) => return format!("[error] Failed to read directory '{}': {}", path_str, e),
-        };
-
-        let mut items: Vec<_> = entries.filter_map(|e| e.ok()).collect();
-        items.sort_by_key(|e| e.file_name());
-
-        let mut lines = Vec::new();
-        for entry in items {
-            let name = entry.file_name().to_string_lossy().to_string();
-            let is_dir = entry.file_type().map(|ft| ft.is_dir()).unwrap_or(false);
-            if is_dir {
-                lines.push(format!("{}/", name));
-            } else {
-                let size = entry.metadata().map(|m| m.len()).unwrap_or(0);
-                lines.push(format!("{} ({}B)", name, size));
-            }
-        }
-
-        if lines.is_empty() {
-            format!("(empty directory: {})", path_str)
-        } else {
-            lines.join("\n")
-        }
-    }
-
-    /// Detect the best available shell on the system.
-    ///
-    /// Priority:
-    /// - Windows: cmd.exe (always available)
-    /// - Unix: $SHELL env var (user's preferred shell) → bash → zsh → sh
-    fn detect_shell() -> (&'static str, &'static str) {
-        #[cfg(windows)]
-        {
-            ("cmd.exe", "/C")
-        }
-
-        #[cfg(not(windows))]
-        {
-            Self::detect_shell_impl(std::env::var("SHELL").ok(), &|s| {
-                std::path::Path::new(s).exists()
-            })
-        }
-    }
-
-    /// Core shell-detection logic with injectable env and filesystem checks
-    /// for testing.
-    ///
-    /// `shell_exists` is a trait object (`&dyn Fn(&str) -> bool`) rather
-    /// than `impl Fn(&str) -> bool` so every caller -- production's real
-    /// `Path::exists` closure and each test's distinct closure -- shares
-    /// exactly ONE monomorphization of this function instead of one per
-    /// closure type (this function was a confirmed generic-monomorphization
-    /// coverage-attribution artifact: every source position had a covered
-    /// instantiation, but the summary table still reported some as missed).
-    #[cfg(not(windows))]
-    fn detect_shell_impl(
-        env_shell: Option<String>,
-        shell_exists: &dyn Fn(&str) -> bool,
-    ) -> (&'static str, &'static str) {
-        if let Some(shell) = env_shell
-            && (shell.ends_with("/zsh") || shell.ends_with("/bash") || shell.ends_with("/sh"))
-            && shell_exists(&shell)
-        {
-            // Only trust `$SHELL` when it actually exists — a stale or
-            // sandbox-missing `$SHELL` (e.g. `/bin/zsh` in an environment that
-            // doesn't ship it) otherwise made every shell call fail to spawn.
-            // When it's missing, fall through to the known-path fallback list.
-            let shell: &'static str = Box::leak(shell.into_boxed_str());
-            return (shell, "-c");
-        }
-        for &shell in &[
-            "/bin/bash",
-            "/usr/bin/bash",
-            "/bin/zsh",
-            "/usr/bin/zsh",
-            "/bin/sh",
-        ] {
-            if shell_exists(shell) {
-                return (shell, "-c");
-            }
-        }
-        ("sh", "-c")
-    }
-
-    async fn shell(&self, args: &Value) -> String {
-        self.shell_with_timeout(args, Duration::from_secs(60)).await
-    }
-
-    /// Same as [`Self::shell`], with an injectable timeout so tests can
-    /// exercise the timeout branch without a real 60-second wait.
-    async fn shell_with_timeout(&self, args: &Value, timeout_duration: Duration) -> String {
-        let command = match args.get("command").and_then(|v| v.as_str()) {
-            Some(c) => c,
-            None => return "[error] missing 'command' argument".to_string(),
-        };
-
-        let workdir = self.ctx.workdir.clone();
-        let (shell, flag) = Self::detect_shell();
-
-        // When a sandbox executor is attached, it builds a command that runs
-        // inside a container / namespace (still targeting `workdir`); otherwise
-        // run the shell directly on the host — the exact prior behavior.
-        let mut cmd = match &self.shell_executor {
-            Some(executor) => executor.build_command(shell, flag, command, &workdir),
-            None => {
-                let mut c = Command::new(shell);
-                c.arg(flag).arg(command).current_dir(&workdir);
-                c
-            }
-        };
-        // Reap the whole command on drop, not just the shell.
-        //
-        // Dropping a `Command` future detaches its process by default, so a
-        // cancelled agent (or an elapsed timeout, which drops the future the
-        // same way) left its shell running: the run vanished from every listing
-        // while its command carried on writing to the workspace. `kill_on_drop`
-        // fixes the shell — but only the shell. Anything the shell itself
-        // started (`sleep 400 && …`) is a *grandchild*, gets reparented to init,
-        // and keeps running. Putting the shell in its own process group and
-        // signalling the group on drop takes the whole tree down with it.
-        cmd.kill_on_drop(true);
-        own_process_group(&mut cmd);
-        // `spawn` inherits stdio where `output` pipes it; pipe explicitly so the
-        // command's output is still captured.
-        cmd.stdout(std::process::Stdio::piped())
-            .stderr(std::process::Stdio::piped());
-
-        // Spawn *inside* the timed future so the reaper guard lives exactly as
-        // long as the command does: dropping this future (timeout, or the whole
-        // batch dropped because the agent was cancelled) drops the guard, which
-        // signals the group. Keeping spawn and wait in one fallible block also
-        // keeps a single error arm, as `Command::output()` had.
-        let run = async {
-            let child = cmd.spawn()?;
-            // The child leads its own group, so its pid is the group id.
-            let _reaper = child.id().map(ProcessGroupReaper);
-            child.wait_with_output().await
-        };
-
-        match timeout(timeout_duration, run).await {
-            Err(_) => format!("[timed out] Command exceeded 60s: {}", command),
-            Ok(Err(e)) => format!("[error] Failed to spawn shell '{}': {}", shell, e),
-            Ok(Ok(output)) => Self::format_command_output(
-                &output.stdout,
-                &output.stderr,
-                output.status.success(),
-                output.status.code().unwrap_or(-1),
-            ),
-        }
-    }
-
-    /// Format captured command output. Split out (behavior-preserving) from
-    /// [`Self::shell_with_timeout`] so the success / non-zero-exit
-    /// stdout+stderr formatting arms can be exercised deterministically on
-    /// every platform, independent of the host shell's command-chaining and
-    /// redirection syntax (`cmd.exe` and `sh` differ, so an integration test
-    /// that produces stdout+stderr+non-zero-exit in one command is not
-    /// portable).
-    fn format_command_output(
-        stdout: &[u8],
-        stderr: &[u8],
-        success: bool,
-        exit_code: i32,
-    ) -> String {
-        let stdout = String::from_utf8_lossy(stdout);
-        let stderr = String::from_utf8_lossy(stderr);
-
-        if success {
-            if stdout.trim().is_empty() {
-                "(command succeeded with no output)".to_string()
-            } else {
-                stdout.to_string()
-            }
-        } else {
-            let mut result = format!("[exit code {}]\n", exit_code);
-            if !stdout.trim().is_empty() {
-                result.push_str(&format!("stdout:\n{}\n", stdout));
-            }
-            if !stderr.trim().is_empty() {
-                result.push_str(&format!("stderr:\n{}", stderr));
-            }
-            result
-        }
     }
 }
